+++ core/tasks.py
@@ -3,7 +3,6 @@ from django_celery_results.models import TaskResult
 from celery import chain
 from app.celery import app
 from core import models as c
-from datasets import models as ds
 from django.conf import settings
 from app.mailer import send_update_error_mail, send_update_success_mail, send_general_task_error_mail
 from core.utils.cache import cache_council_property_summaries_full, cache_community_property_summaries_full
@@ -93,6 +92,8 @@ def async_create_update(self, dataset_id, file_id=None):
 
 @app.task(bind=True, queue='celery', acks_late=True, max_retries=1)
 def async_annotate_properties_with_all_datasets(self):
+    from datasets import models as ds
+
     for model_name in ds.PropertyAnnotation.ANNOTATED_DATASETS:
         async_annotate_properties_with_dataset.delay(getattr(ds, model_name).get_dataset().id)